Following the ducks pattern now
