+++ src/client/store/index.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware, compose } from "redux";
 import createSagaMiddleware from "redux-saga";
-import rootReducer from "./reducers";
+import rootReducer from "./rootReducer";
 import sagas from "./sagas";
 
 const sagaMiddleware = createSagaMiddleware();
