+++ src/client/store/modules/app.js
@@ -1,4 +1,12 @@
-import { START_LOADING_APP, STOP_LOADING_APP } from "state/actions/types";
+import { START_LOADING_APP, STOP_LOADING_APP } from "store/actionTypes";
+
+export const startLoadingApp = () => ({
+  type: START_LOADING_APP
+});
+
+export const stopLoadingApp = () => ({
+  type: STOP_LOADING_APP
+});
 
 export default (
   state = {
