+++ src/client/store/modules/topic.js
@@ -5,7 +5,33 @@ import {
   STOP_LOADING_TOPICS,
   SET_CURRENT_TOPIC,
   EMPTY_CURRENT_TOPIC
-} from "state/actions/types";
+} from "store/actionTypes";
+
+export const addTopics = topics => ({
+  type: ADD_TOPICS,
+  topics
+});
+
+export const emptyTopics = () => ({
+  type: EMPTY_TOPICS
+});
+
+export const startLoadingTopics = () => ({
+  type: START_LOADING_TOPICS
+});
+
+export const stopLoadingTopics = () => ({
+  type: STOP_LOADING_TOPICS
+});
+
+export const setCurrentTopic = topic => ({
+  type: SET_CURRENT_TOPIC,
+  topic
+});
+
+export const emptyCurrentTopic = () => ({
+  type: EMPTY_CURRENT_TOPIC
+});
 
 export default (
   state = {
